requirements.txt
networkx==3.1
pyvis==0.3.2
dash==2.14.2
dash-cytoscape==0.3.0
dash-bootstrap-components==1.5.0
ast-comments==1.0.1
typing-extensions==4.8.0
gitpython==3.1.40
tqdm==4.66.1
joblib==1.3.2 